書式指定子とポインタ型の不一致を直す
%08x や %u にポインタ型の値をそのまま渡していた。UINT_PTR へ明示的に変換する。
gen-syms.cc は time_t を %d に渡していたので、出力先の dump_version が int で
あることに合わせて変換する。

+++ src/gen-syms.cc
@@ -3049,7 +3049,7 @@ process_interactive ()
 static void
 print_version ()
 {
-  printf ("int dump_version = %d;\n", time (0));
+  printf ("int dump_version = %d;\n", static_cast <int> (time (0)));
 }
 
 void
